+++ src/user-lists/user-list.controller.ts
@@ -10,6 +10,7 @@ import {
   ApiParam,
 } from "@nestjs/swagger";
 
+import { DbUserHighlight } from "../user/entities/user-highlight.entity";
 import { DbUser } from "../user/user.entity";
 import { PageOptionsDto } from "../common/dtos/page-options.dto";
 import { ApiPaginatedResponse } from "../common/decorators/api-paginated-response.decorator";
@@ -17,6 +18,7 @@ import { PageDto } from "../common/dtos/page.dto";
 import { UserId } from "../auth/supabase.user.decorator";
 import { SupabaseGuard } from "../auth/supabase.guard";
 
+import { HighlightOptionsDto } from "../highlight/dtos/highlight-options.dto";
 import { CreateUserListDto } from "./dtos/create-user-list.dto";
 import { DbUserList } from "./entities/user-list.entity";
 import { UserListService } from "./user-list.service";
@@ -45,6 +47,16 @@ export class UserListController {
     @Query() pageOptionsDto: PageOptionsDto
   ): Promise<PageDto<DbUserList>> {
     return this.userListService.findAllByUserId(pageOptionsDto, userId);
+  }
+
+  @Get("/featured")
+  @ApiOperation({
+    operationId: "getFeaturedLists",
+    summary: "Gets public featured lists",
+  })
+  @ApiOkResponse({ type: DbUserList })
+  async getFeaturedLists(@Query() pageOptionsDto: PageOptionsDto): Promise<PageDto<DbUserList>> {
+    return this.userListService.findAllFeatured(pageOptionsDto);
   }
 
   @Post("/")
@@ -197,6 +209,23 @@ export class UserListController {
     @Param("userListContributorId") userListContributorId: string
   ): Promise<void> {
     await this.userListService.deleteUserListContributor(id, userListContributorId);
+  }
+
+  @Get("/:id/contributors/highlights")
+  @ApiOperation({
+    operationId: "getUserListContributorHighlights",
+    summary: "Retrieves highlights for contributors for an individual user list",
+  })
+  @ApiPaginatedResponse(DbUserHighlight)
+  @ApiOkResponse({ type: DbUserHighlight })
+  @ApiNotFoundResponse({ description: "Unable to get user list contributor highlights" })
+  @ApiBadRequestResponse({ description: "Invalid request" })
+  @ApiParam({ name: "id", type: "string" })
+  async getUserListContributorHighlights(
+    @Query() pageOptionsDto: HighlightOptionsDto,
+    @Param("id") id: string
+  ): Promise<PageDto<DbUserHighlight>> {
+    return this.userListService.findListContributorsHighlights(pageOptionsDto, id);
   }
 
   @Get("/timezones")